next: Build all vendor variants of interface libs
With Android 12, the libs need to be specified explicitly to be built into
the vendor as they are no longer part of the VNDK.

This is because the VNDK is becoming bigger with every release
and needs to be trimmed down.

Some libs are still built by default, but it doesn't hurt to have
a build specification for vendor builds.

Change-Id: Ic8fddbb9341385ee66b22a93a807eed944a6a9b1